Initial version of fixed_hashmap
Added fixed_hashmap, a fixed replacement for unordered_map. It is the first version while I figure out the details for the more general hashmap and eventually the hashset which will be a replacement for unordered_set
Added unit tests for the different functionality and comments to explain where the interface differs from the STL
Eventually we'll start moving more functionality into hashmap_common as we extract the common parts
This hashmap was compared against the robin_hood hashmap, ankerl::unordered_dense, std::unordered_map and eastl::hash_map inserting and erasing random integers. More tests are needed to evaluate performance, especially as the general-purpose hashmap will need to make decisions regarding memory allocation and rehashing
However, as far as I can tell, many of these hashmaps don't benefit much from reserving memory, and I think we can
In any case, this is a promising avenue that will continue to be explored