Improve bitset codegen
It was a bit of a rabbit hole of trying out a base class and template instantiation, if constexpr and looking at loop unrolling. In the end I decided against the complexity and instead rewrote the functions that are problematic even in optimized builds which are prevented from unrolling due to early outs. As the common case for bitset is small, we cater for those